chore to clean and move funcs to proper module
